LayoutTests/inspector/timeline/timeline-event-EventDispatch.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/timeline-event-utilities.js"></script>
<script>
function testClickEventHandler({preventDefault}) {
    let button = document.body.appendChild(document.createElement("button"));
    button.addEventListener("click", (event) => {
        TestPage.addResult("click handler fired" + (preventDefault ? ", will prevent default" : ""));
        setTimeout(() => { finishRecording({preventDefault}); });
        if (preventDefault)
            event.preventDefault();
    });

    button.dispatchEvent(new MouseEvent("click", {bubbles: true, cancelable: true}));
}

function testClickEventAttributeHandler(id, {preventDefault}) {
    let button = document.getElementById(id);
    button.dispatchEvent(new MouseEvent("click", {bubbles: true, cancelable: true}));
    setTimeout(() => { finishRecording({preventDefault}); });
}

function test()
{
    let suite = InspectorTest.createAsyncSuite("TimelineEvent.EventDispatch");

    suite.addTestCase({
        name: "TimelineEvent.EventDispatch.Handler.Regular",
        test(resolve, reject) {
            captureTimelineWithScript(`testClickEventHandler({preventDefault: false})`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.preventDefault === "boolean");

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.EventDispatched);
                InspectorTest.expectEqual(records.length, 1, "Should be 1 EventDispatched record.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + JSON.stringify(record.extraDetails));
                    InspectorTest.expectEqual(record.extraDetails.defaultPrevented, pageRecordingData.preventDefault, "ScriptTimelineRecord extraDetails should show default was not prevented.");
                }
            }).then(resolve, reject);
        }
    });

    suite.addTestCase({
        name: "TimelineEvent.EventDispatch.Handler.DefaultPrevented",
        test(resolve, reject) {
            captureTimelineWithScript(`testClickEventHandler({preventDefault: true})`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.preventDefault === "boolean");

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.EventDispatched);
                InspectorTest.expectEqual(records.length, 1, "Should be 1 EventDispatched record.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + JSON.stringify(record.extraDetails));
                    InspectorTest.expectEqual(record.extraDetails.defaultPrevented, pageRecordingData.preventDefault, "ScriptTimelineRecord extraDetails should show default was prevented.");
                }
            }).then(resolve, reject);
        }
    });

    suite.addTestCase({
        name: "TimelineEvent.EventDispatch.AttributeHandler.Regular",
        test(resolve, reject) {
            captureTimelineWithScript(`testClickEventAttributeHandler("b1", {preventDefault: false})`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.preventDefault === "boolean");

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.EventDispatched);
                InspectorTest.expectEqual(records.length, 1, "Should be 1 EventDispatched record.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + JSON.stringify(record.extraDetails));
                    InspectorTest.expectEqual(record.extraDetails.defaultPrevented, pageRecordingData.preventDefault, "ScriptTimelineRecord extraDetails should show default was not prevented.");
                }
            }).then(resolve, reject);
        }
    });

    suite.addTestCase({
        name: "TimelineEvent.EventDispatch.AttributeHandler.DefaultPrevented",
        test(resolve, reject) {
            captureTimelineWithScript(`testClickEventAttributeHandler("b2", {preventDefault: true})`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.preventDefault === "boolean");

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.EventDispatched);
                InspectorTest.expectEqual(records.length, 1, "Should be 1 EventDispatched record.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + JSON.stringify(record.extraDetails));
                    InspectorTest.expectEqual(record.extraDetails.defaultPrevented, pageRecordingData.preventDefault, "ScriptTimelineRecord extraDetails should show default was prevented.");
                }
            }).then(resolve, reject);
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Tests 'EventDispatch' Timeline event records.</p>
<button id="b1" onclick="TestPage.addResult('b1 onclick attribute handler fired'); return true"></button>
<button id="b2" onclick="TestPage.addResult('b2 onclick attribute handler fired, will prevent default'); return false"></button>
</body>
</html>
