LayoutTests/inspector/timeline/timeline-event-EventDispatch-expected.txt

Tests 'EventDispatch' Timeline event records.

 

== Running test suite: TimelineEvent.EventDispatch
-- Running test case: TimelineEvent.EventDispatch.Handler.Regular
Starting Capture...
click handler fired
Stopping Capture...
PASS: Should be 1 EventDispatched record.
DETAILS: {"type":"click","defaultPrevented":false}
PASS: ScriptTimelineRecord extraDetails should show default was not prevented.

-- Running test case: TimelineEvent.EventDispatch.Handler.DefaultPrevented
Starting Capture...
click handler fired, will prevent default
Stopping Capture...
PASS: Should be 1 EventDispatched record.
DETAILS: {"type":"click","defaultPrevented":true}
PASS: ScriptTimelineRecord extraDetails should show default was prevented.

-- Running test case: TimelineEvent.EventDispatch.AttributeHandler.Regular
Starting Capture...
b1 onclick attribute handler fired
Stopping Capture...
PASS: Should be 1 EventDispatched record.
DETAILS: {"type":"click","defaultPrevented":false}
PASS: ScriptTimelineRecord extraDetails should show default was not prevented.

-- Running test case: TimelineEvent.EventDispatch.AttributeHandler.DefaultPrevented
Starting Capture...
b2 onclick attribute handler fired, will prevent default
Stopping Capture...
PASS: Should be 1 EventDispatched record.
DETAILS: {"type":"click","defaultPrevented":true}
PASS: ScriptTimelineRecord extraDetails should show default was prevented.
